Fixed gitignore , added engineer and marketplace , fixed bugs in registeration and login, docker fix to deploy in aws

+++ Domain/Models/User/ServiceProvider.cs
@@ -17,6 +17,6 @@ namespace EgyptOnline.Models
 
         public string ProviderType { get; set; } = "Worker";
 
-        public int ServicePricePerDay { get; set; } = 0;
+        public decimal ServicePricePerDay { get; set; } = 0;
     }
 }
